Better island gen. Better FOV.
Chrome's profiler reports that ROT's "toHex" function
is massively slower than its "toRGB" function, so switch.

Don't compute bresenhamTranslucence() for tiles that are
definitely fully visible (i.e. are not even partially in the
shadow of any not-fully-translucent tiles).

Add a couple of manmade structures to the island; but this
isn't really good walls-and-fences gen yet.